+++ src/main/java/com/inter/SistemaDeCadastro/controllers/dtos/EscolaridadeDto.java
@@ -3,7 +3,7 @@ package com.inter.SistemaDeCadastro.controllers.dtos;
 import jakarta.validation.constraints.NotBlank;
 
 public record EscolaridadeDto(
-        Long codEscolaridade,
+        Integer codEscolaridade,
         @NotBlank String nome,
         String status
 ) {
