+++ src/main/java/com/inter/SistemaDeCadastro/interfaces/FamiliarRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FamiliarRepository extends JpaRepository<FamiliarModel,Long> {
-    List<FamiliarModel> findByAlunoCodAluno(Long codAluno);
+public interface FamiliarRepository extends JpaRepository<FamiliarModel,Integer> {
+    List<FamiliarModel> findByAlunoCodAluno(Integer codAluno);
     List<FamiliarModel> findByStatus(String status);
 }
