+++ src/main/java/com/inter/SistemaDeCadastro/interfaces/OcupacaoRepository.java
@@ -3,5 +3,5 @@ package com.inter.SistemaDeCadastro.interfaces;
 import com.inter.SistemaDeCadastro.models.OcupacaoModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OcupacaoRepository extends JpaRepository<OcupacaoModel,Long> {
+public interface OcupacaoRepository extends JpaRepository<OcupacaoModel,Integer> {
 }
