+++ src/main/java/com/inter/SistemaDeCadastro/interfaces/RoleRepository.java
@@ -3,6 +3,6 @@ package com.inter.SistemaDeCadastro.interfaces;
 import com.inter.SistemaDeCadastro.models.RoleModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<RoleModel,Long> {
+public interface RoleRepository extends JpaRepository<RoleModel,Integer> {
     RoleModel findByNome(String nome);
 }
