+++ src/main/java/com/inter/SistemaDeCadastro/interfaces/TurnoRepository.java
@@ -3,5 +3,5 @@ package com.inter.SistemaDeCadastro.interfaces;
 import com.inter.SistemaDeCadastro.models.TurnoModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TurnoRepository extends JpaRepository<TurnoModel,Long> {
+public interface TurnoRepository extends JpaRepository<TurnoModel,Integer> {
 }
